pkg/tmux: drop redundant int64 wrap on maxAgeMillis
math.MaxInt64 divided by int64(time.Millisecond) is already an int64
constant; the outer int64() conversion was an identity no-op. The
inner conversion stays to force integer division.

+++ pkg/tmux/notification.go
@@ -152,7 +152,7 @@ func (n Notification) ExtendedOutput() (ExtendedOutputNotification, bool, error)
 
 // maxAgeMillis is the largest %extended-output age in milliseconds that fits in
 // a time.Duration without overflow.
-const maxAgeMillis = int64(math.MaxInt64 / int64(time.Millisecond))
+const maxAgeMillis = math.MaxInt64 / int64(time.Millisecond)
 
 // SubscriptionChanged returns the typed form of a `%subscription-changed` notification.
 func (n Notification) SubscriptionChanged() (SubscriptionChangedNotification, bool, error) {
